+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
 # Next Release
-
+- [#395] (https://github.com/IAMconsortium/pyam/pull/403_ - Add boxplot function
 - [#394](https://github.com/IAMconsortium/pyam/pull/394) Switch CI to Github Actions.
 - [#393](https://github.com/IAMconsortium/pyam/pull/393) Import ABC from collections.abc for Python 3.10 compatibility.
 - [#380](https://github.com/IAMconsortium/pyam/pull/380) Add compatibility with latest matplotlib and py3.8